Refacto OpenMessagePromptContext Externalise in a new componant.
