+++ frontend/src/components/templates/Form/HideSectionsContainer.tsx
@@ -1,4 +1,4 @@
-import React, {
+import {
   Children,
   cloneElement,
   createContext,
